kernel: removed pid argument from rd/wr/sd/recv.
This was only used for statistical purposes (input/output). Passing the
pid as an argument to all kinds of functions just for this purpose is
not a good idea. Instead, we do the statistics in the syscall functions
now. This changes the semantics of input/output stats slightly, because
it is now only the input/output directly caused by the process via
read/write/send/receive system calls. ELF loading, demand loading,
messages from e.g. delegate and so on, are no longer counted.

+++ source/kernel/include/vfs/device.h
@@ -92,8 +92,8 @@ public:
 	/**
 	 * Sends the given message to the channel <chan>, which belongs to this device.
 	 */
-	ssize_t send(VFSChannel *chan,ushort flags,msgid_t id,USER const void *data1,
-                 size_t size1,USER const void *data2,size_t size2);
+	int send(VFSChannel *chan,ushort flags,msgid_t id,USER const void *data1,
+             size_t size1,USER const void *data2,size_t size2);
 
 	/**
 	 * Receives a message from the channel <chan>, which belongs to this device.
